refactor: 프로그래머스 181187 코드 리팩토링

+++ MyungJi/240101_week/programmers_181187_두원사이의정수쌍.cpp
@@ -1,5 +1,5 @@
 #include <bits/stdc++.h>
-long long maxY, minY;
+int maxY, minY;
 using namespace std;
 
 long long solution(int r1, int r2) {
